refactor createEntries
